server/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.7.0
python-dotenv==1.0.0
asgiref==3.9.1
sqlparse==0.5.3
